# General improvements to triple moving average