I've fiddled around with `triple_moving_average.py` file and did some improvements.

* Implemented a simple "timing" for benchmarking the filtering process
* Reduced execution time by two orders of magnitude by using `enumerate`
	* I noticed that there were many calls to `.index()` in `triple_moving_average` function. Since numpy probably uses linear search to find the index,  it generally boils down to factorial complexity if  `.index()` is called in a loop.
	* By refactoring it with `for index, point in enumerate(signal_array)` we get both the index of the point and the point itself, with linear complexity.
	* This lead to reducing execution time **from ~6 seconds to ~0.08 seconds** (on my machine) using orbit.csv as the dataset and window size of 2.
	* With this performance improvement, we can use silly window sizes (f.eg. 20) and still perform way quicker than using size of 2 with code prior to refactor (~0.4 seconds with size 20 vs ~6 seconds with size 2).
	    * Before (size=2, ~6 seconds)
               ![](https://i.imgur.com/jbKzzWI.png)
            * After (size=20, ~0.4 seconds)
              ![](https://i.imgur.com/QcfUAGr.png)
* Docstring for function `generate_filtered_data` implied that input was a file, not a numpy array - fix